Fix responsive layout
