+++ src/Components/Inspirations/Inspirations.scss
@@ -71,8 +71,11 @@ $arrow-base-size: $button-size/5;
   display: flex;
   flex-direction: row;
   align-items: center;
-  justify-content: space-evenly;
+  justify-content: space-around;
   margin-bottom: 1em;
+  height: 100%;
+  overflow-x: hidden;
+  overflow: hidden;
 }
 
 .img-inspo {
@@ -89,7 +92,7 @@ $arrow-base-size: $button-size/5;
   height: 6.5vh;
   justify-content: center;
   flex-direction: column;
-  margin-left: 1em;
+  margin-left: 1.5em;
 }
 
 .image-box, .word-box {
@@ -115,19 +118,34 @@ $arrow-base-size: $button-size/5;
   border-radius: 3px;
 }
 
+.choose-time-btn {
+  background-color: #3b544d;
+  color: white;
+  width: 100%;
+  height: 3vh;
+  box-shadow: none;
+  border: none;
+  border-radius: 3px;
+  margin-top: 2em;
+}
+
 @media screen and (min-width: 40em) {
+  .container {
+    display: flex;
+    flex-direction: column;
+    width: 76%;
+    margin: 0;
+  }
   .inspirations {
     display: flex;
     flex-direction: column;
-    justify-content: center;
-    height: 100%;
     margin-left: 2em;
     align-items: flex-start;
+    width: 60%;
   }
   .image-box, .word-box {
     height: 10%;
     width: 30%;
-    margin-top: 10%;
   }
   .play-arrow-btn {
     border-width: $arrow-base-size*2 0 $arrow-base-size*2 ($arrow-base-size*1.6)*2;
@@ -141,6 +159,7 @@ $arrow-base-size: $button-size/5;
   .word-inspo {
     font-size: 32px;
     margin-left: 2vw;
+    margin-top: 45px;
   }
   .audio-box {
     display: flex;
@@ -157,6 +176,10 @@ $arrow-base-size: $button-size/5;
     display: flex;
     justify-content: center;
     justify-content: flex-start;
+  }
+  .choose-time-btn{
+    width: 15em;
+    margin-left: 2em;
   }
 }
