+++ src/Components/TextInput/TextInput.scss
@@ -1,8 +1,9 @@
 textarea {
-    width: 100%;
-    height: 35vh;
-    background-color: #F8E9B3;
-    font-size: 1em;
+  width: 98%;
+  height: 28vh;
+  background-color: #F8E9B3;
+  font-size: 1em;
+  overflow: hidden;
 }
 
 form {
@@ -30,7 +31,7 @@ form > button {
     justify-content: center;
   }
   textarea {
-    width: 100%;
+    width: 70vw;
     height: 74vh;
     margin-top: 4%;
   }
