+++ src/Components/Timer/Timer.scss
@@ -4,6 +4,7 @@
   width: 5rem;
   background: none;
   cursor: pointer;
+  overflow: hidden;
 }
 
 .start-stop-btn {
