+++ src/Components/WritingPage/WritingPage.scss
@@ -2,15 +2,15 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
+  overflow-x: hidden;
+  overflow: hidden;
 }
 
 @media screen and (min-width: 40em) {
-  .writing-inspiration {
+  .writing-page {
     display: flex;
     flex-direction: row;
     justify-content: center;
-    height: 100%;
-    width: 100%;
   }
 }
